Update GoogleCast and GoogleInteractiveMediaAds dependencies
Update GoogleCast and GoogleInteractiveMediaAds dependencies

+++ Frameworks/GoogleInteractiveMediaAds.framework/Headers/IMAAdPodInfo.h
@@ -16,12 +16,12 @@
 /**
  *  Total number of ads in the pod this ad belongs to. Will be 1 for standalone ads.
  */
-@property(nonatomic, readonly) int totalAds;
+@property(nonatomic, readonly) NSInteger totalAds;
 
 /**
  *  The position of this ad within an ad pod. Will be 1 for standalone ads.
  */
-@property(nonatomic, readonly) int adPosition;
+@property(nonatomic, readonly) NSInteger adPosition;
 
 /**
  *  Specifies whether the ad is a bumper. Bumpers are short videos used to open
@@ -30,10 +30,16 @@
 @property(nonatomic, readonly) BOOL isBumper;
 
 /**
- *  The index of the pod, where pre-roll pod is 0, mid-roll pods are 1 .. N
- *  and the post-roll is -1.
+ *  Client side: Returns the index of the ad pod. For a preroll pod, returns 0.
+ *  For midrolls, returns 1, 2,..., N. For a postroll pod, returns -1. Defaults
+ *  to 0 if this ad is not part of a pod, or this pod is not part of a playlist.
+ *  DAI live stream: Always returns -1.
+ *  DAI VOD: Returns the index of the ad pod. For a preroll pod, returns 0. For
+ *  midrolls, returns 1, 2,...,N. For a postroll pod, returns N+1...N+X.
+ *  Defaults to 0 if this ad is not part of a pod, or this pod is not part of a
+ *  playlist.
  */
-@property(nonatomic, readonly) int podIndex;
+@property(nonatomic, readonly) NSInteger podIndex;
 
 /**
  *  The position of the pod in the content in seconds. Pre-roll returns 0,
@@ -46,6 +52,9 @@
  */
 @property(nonatomic, readonly) NSTimeInterval maxDuration;
 
+/**
+ * :nodoc:
+ */
 - (instancetype)init NS_UNAVAILABLE;
 
 @end
